+++ src/wai/common/adams/imaging/locateobjects/__init__.py
@@ -3,3 +3,6 @@ Package for working with located objects in image classification/identification 
 """
 from ._LocatedObject import LocatedObject
 from ._LocatedObjects import LocatedObjects
+from ._NormalizedLocatedObject import NormalizedLocatedObject
+from ._NormalizedLocatedObjects import NormalizedLocatedObjects
+from ._convert import absolute_to_normalized, normalized_to_absolute