+++ src/wai/common/geometry/__init__.py
@@ -4,3 +4,6 @@ Package for working with shapes.
 from ._Point import Point
 from ._Polygon import Polygon
 from ._Rectangle import Rectangle
+from ._NormalizedPoint import NormalizedPoint
+from ._NormalizedPolygon import NormalizedPolygon
+from ._NormalizedRectangle import NormalizedRectangle